inference/requirements.txt
aio_pika==9.0.5
bentoml==1.0.16
fastapi==0.95.0
loguru==0.6.0
nest_asyncio==1.5.6
onnx==1.13.1
onnxruntime-gpu==1.14.1
pillow==9.5.0
Pillow==9.5.0
pyarrow==11.0.0
wget==3.2
